refactor(ui): package menus and move grant/feat formatting to core
Reorganize hub/creation/progression/scenario packages, extract
core.grants.format and core.feats.requirement_text, thin UI display.
