+++ ui/menus/__init__.py
@@ -1,12 +1,12 @@
 """Публичный API экранов меню."""
 
-from ui.menus._creation_steps import show_create_character_flow
-from ui.menus.characters_menu import show_characters_menu
-from ui.menus.load_game import show_load_game_flow
-from ui.menus.main_menu import show_main_menu, show_welcome_screen
-from ui.menus.mods_menu import show_mods_menu
-from ui.menus.new_game import show_new_game_flow
-from ui.menus.settings import (
+from ui.menus.creation.steps import show_create_character_flow
+from ui.menus.hub.characters_menu import show_characters_menu
+from ui.menus.hub.load_game import show_load_game_flow
+from ui.menus.hub.main_menu import show_main_menu, show_welcome_screen
+from ui.menus.hub.mods_menu import show_mods_menu
+from ui.menus.hub.new_game import show_new_game_flow
+from ui.menus.hub.settings import (
     select_difficulty,
     show_languages_menu,
     show_settings,
