+++ ui/menus/creation/__init__.py
@@ -1,0 +1,1 @@
+"""UI пакет creation."""
