+++ ui/menus/creation/equipment.py
