+++ ui/menus/creation/navigation.py
@@ -3,7 +3,7 @@
 from core.feats.catalog import race_feat_step_required
 from core.mechanics.expertise import expertise_step_required
 from core.progression.class_progression import subclass_offered_at_creation
-from ui.menus._creation_state import CreationStep, _CreationState
+from ui.menus.creation.state import CreationStep, _CreationState
 
 
 def feats_step_required(state: _CreationState) -> bool:
