+++ ui/menus/hub/__init__.py
@@ -1,0 +1,1 @@
+"""UI пакет hub."""
