+++ ui/menus/progression/__init__.py
@@ -1,0 +1,1 @@
+"""UI пакет progression."""
