+++ ui/menus/scenario/__init__.py
@@ -1,0 +1,1 @@
+"""UI пакет scenario."""
